UIDEXP-55: Move MappingProfilesContainer under MappingProfiles namespace
